Use lru_cache from methodtools on methods
instead of functools.lru_cache. The new variant creates a separate cache
for each object, while functools.lru_cache keeps a common cache for all
instances of the class. Keeping separate caches is preferable in the
uses introduced so far.
Also refactor so that we never use @property and @methodtools.lru_cache
on the same function, so that we can introspect the lru_cache by calling
cache_info(). This data can be used to tweak the cache sizes.